trying to clean this up a bit
