--- app/tracks/layout.tsx
@@ -1,11 +1,0 @@
-import TrackTitle from "./title/Title";
-
-const Tracks = () => {
-  return (
-    <div>
-      <TrackTitle />
-    </div>
-  );
-};
-
-export default Tracks;
